+++ lib/new_relic/agent/instrumentation/elasticsearch/instrumentation.rb
@@ -52,11 +52,13 @@ module NewRelic::Agent::Instrumentation
     end
 
     def nr_cluster_name
-      return @nr_cluster_name if @nr_cluster_name
+      return @nr_cluster_name if defined?(@nr_cluster_name)
       return if nr_hosts.empty?
 
       NewRelic::Agent.disable_all_tracing do
-        @nr_cluster_name ||= perform_request('GET', '_cluster/health').body['cluster_name']
+        @nr_cluster_name ||= if NewRelic::Agent.config[:'elasticsearch.perform_health_check']
+          perform_request('GET', '_cluster/health').body['cluster_name']
+        end
       end
     rescue StandardError => e
       NewRelic::Agent.logger.error('Failed to get cluster name for elasticsearch', e)